visitor classes refactored so they're more versatile (fixes: #5)
The previous version depended on method names which didn't leave
much room to encode much useful information for deciding which
method to use. I.e. classes could only be used as a base as
long as the class wasn't named Settings.RecursiveContainer or
something similar. Likewise we coudn't properly define a visitor
for XML tags like some-thing.

Well, we could but that required using hexadecimals in a name
and that is just plain wrong.