added scheme for config version compatibility handling (fixes: #4)
If the config version is bumped, but the version should stay
compatible, we need to upgrate the config so it can be validated
with a newer version. So if the current config version is 1.1 and
the loaded is 1.0, we'll update from 1.0 to 1.1 . If a new version
is added, then 1.0 will be updated to 1.1 and then to 1.2.