copy generators before binding them to avoid problems from using a generator twice in a config
If the same object generator with a list of subgenerators is used
more than once the bind() method will be called twice on the same
instance (the decorated functions which became sub-generators) so
copy the generator on registering. This decreases efficiency  but
this  way we have a guarantee that the instance is "fresh" before
calling bind().