generators changed so they don't have matcher/priority attributes and the bind method
Since these attributes were only there for the sake of the
PriorityGenerator it wasn't really appropriate for them to
be completely bound to the generator sub-system. So they've
been removed and PriorityGenerator handles these on registration.

The bind() method has been removed since it was counter-productive.
Each time a generator was used the bind method was called. Since a
generator can only be bound once and this was done on instances, the
result was that two subsequent usages raised a problem. This was fixed,
but the fix involved copying the generator, which isn't very nice.

To achieve this I've introduced three new decorators in
codega.generator:

 * generator - This will mark the function as a generator function.
   The factory method, if given will be called on func when the
   generator is registered
 * match - Add a matcher to the generator.
 * priority - Set the priority of the generator.