change ply.yacc wrapper to never write a parser tab file
codega.alp.parser.ParserBase() used to specify a tempfile to yacc.yacc()
as the tabmodule to be used, however yacc() expects an actual module name
there which can be imported by python itself, using an exec("import %s")
call.

If we are using a filename (which codega was doing), the import statement
will fail.

This patch changes the yacc invocation to never generate the
tabfile instead, which is functionally equivalent.

Change-Id: Ia6f761b39f31c605bd50f7444dbe2f0a10cae3b9
Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>